+++ android/Bottari/presentation/src/main/java/com/bottari/presentation/view/common/decoration/ItemSpacingDecoration.kt
@@ -1,4 +1,4 @@
-package com.bottari.presentation.common
+package com.bottari.presentation.view.common.decoration
 
 import android.graphics.Rect
 import android.view.View
